# backend/src/main/resources/static/swagger-ui/openapi3.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API 문서",
    "description" : "RestDocsWithSwagger Docs",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:8081/api"
  } ],
  "tags" : [ ],
  "paths" : {
    "/agenda" : {
      "put" : {
        "tags" : [ "agenda" ],
        "operationId" : "update-agenda-assignee",
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/agenda-1829758783"
              },
              "examples" : {
                "update-agenda-assignee" : {
                  "value" : "{\r\n  \"agendaId\" : 1,\r\n  \"content\" : \"수정된 내용\",\r\n  \"status\" : \"완료\"\r\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "200"
          }
        }
      },
      "post" : {
        "tags" : [ "agenda" ],
        "operationId" : "agenda-savesave-agenda",
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/agenda1755123133"
              },
              "examples" : {
                "agenda-save" : {
                  "value" : "{\r\n  \"issueId\" : 1,\r\n  \"content\" : \"의사결정 제공\"\r\n}"
                },
                "save-agenda" : {
                  "value" : "{\r\n  \"issueId\" : 1,\r\n  \"content\" : \"의사결정 제공\"\r\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "201"
          }
        }
      }
    },
    "/agenda/1" : {
      "delete" : {
        "tags" : [ "agenda" ],
        "operationId" : "delete-agenda-assignee",
        "responses" : {
          "200" : {
            "description" : "200"
          }
        }
      }
    },
    "/agenda/assignee" : {
      "post" : {
        "tags" : [ "agenda" ],
        "operationId" : "save-agenda-assignee",
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/agenda-assignee-831680420"
              },
              "examples" : {
                "save-agenda-assignee" : {
                  "value" : "{\r\n  \"agendaId\" : 1,\r\n  \"assigneeId\" : 1\r\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "201"
          }
        }
      }
    },
    "/hello" : {
      "get" : {
        "tags" : [ "hello" ],
        "operationId" : "hello-get",
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/hello-1270671984"
                },
                "examples" : {
                  "hello-get" : {
                    "value" : "{\r\n  \"body\" : \"hello\"\r\n}"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "hello" ],
        "operationId" : "hello-post",
        "responses" : {
          "201" : {
            "description" : "201"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "agenda1755123133" : {
        "type" : "object",
        "properties" : {
          "issueId" : {
            "type" : "number",
            "description" : "이슈 Id"
          },
          "content" : {
            "type" : "string",
            "description" : "의사결정 내용"
          }
        }
      },
      "agenda-1829758783" : {
        "type" : "object",
        "properties" : {
          "agendaId" : {
            "type" : "number",
            "description" : "의사결정 Id"
          },
          "content" : {
            "type" : "string",
            "description" : "수정할 내용"
          },
          "status" : {
            "type" : "string",
            "description" : "상태 변경,[ASSIGNED,PROGRESS,COMPLETED,APPROVED]"
          }
        }
      },
      "agenda-assignee-831680420" : {
        "type" : "object",
        "properties" : {
          "assigneeId" : {
            "type" : "number",
            "description" : "담당자 Id"
          },
          "agendaId" : {
            "type" : "number",
            "description" : "의사결정 Id"
          }
        }
      },
      "hello-1270671984" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string",
            "description" : "바디"
          }
        }
      }
    }
  }
}